data handling: removes hacks/createCachedSelector
This previously (probably up to reselect 5.*) was necessary to not have
data fetched by rtk query to be GC'ed when only accessed through
(cached?) selectors. Since then the memoization function has changed and
seems to have fixed the performance issues when not using
createCachedSelector and the garbage collection of actively used data.